fix tests when run on not MongoDb enviornment
