+++ lib/utils/express/src/index.js
@@ -26,7 +26,7 @@ const vcapenv = require('abacus-vcapenv');
 const rc = require('abacus-rc');
 const oauth = require('abacus-oauth');
 const inflight = require('./lib/middleware/inflight');
-const createRateLimiter = require('./lib/middleware/rate-limiting/rate-limiter');
+const createRateLimiter = require('./lib/middleware/rate-limiter/mongo-rate-limiter');
 const oauthContext = require('./lib/middleware/oauth-context');
