+++ lib/utils/express/src/test/middleware/rate-limiting/mongo-store.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const moment = require('abacus-moment');
-const MongoStore = require('../../../lib/middleware/rate-limiting/mongo-store');
+const MongoStore = require('../../../lib/middleware/rate-limiter/mongo-store');
 
 describe('MongoStore tests', () => {
   const firstRequestTimestamp = 42;
